requirements.txt
google-generativeai
pandas
streamlit
scikit-learn==1.2.2
deep_translator
smtplib
email